requirements.txt
opencv-python
pandas
openpyxl
qrcode[pil]
Pillow